+++ tests/unit/src/games/blackjack_tests/blackjack_player_test.py
@@ -4,7 +4,6 @@
 #	resultay | 14-09-23 | Initial version
 
 import pytest
-from src.general import Face
 from src.games.blackjack.constants import BLACKJACK
 from src.games.blackjack.constants import BlackjackError
 from src.games.blackjack.blackjack_player import BlackjackPlayer
@@ -17,78 +16,6 @@ def player() -> BlackjackPlayer:
 
 class TestDoubleDown():
     """test double down related functions"""
-    def should_double_down(
-        self,
-        card: Card,
-        mask: list[int],
-    ) -> None:
-        """test_should_double_down helper"""
-        player = BlackjackPlayer('David Blaine')
-        player.add_card(card)
-        for i in range(2, 12):
-            if i in mask:
-                assert player.should_double_down(i)
-            else:
-                assert not player.should_double_down(i)
-
-    @pytest.mark.parametrize(
-        'points, mask',
-        [
-            (13, [5, 6]),
-            (14, [5, 6]),
-            (15, range(4, 7)),
-            (16, range(4, 7)),
-            (17, range(3, 7)),
-            (18, range(2, 7)),
-            (19, [6]),
-            (20, []),
-            (21, []),
-        ]
-    )
-    def test_should_double_down_soft(
-        self,
-        ace: Card,
-        mask: list[int],
-        points: int,
-    ):
-        """player should double down with soft hand"""
-        # Soft 19 doubles against dealer 6
-        # Soft 18 doubles against dealer 2 through 6
-        # Soft 17 doubles against dealer 3 through 6
-        # Soft 16 doubles against dealer 4 through 6
-        # Soft 15 doubles against dealer 4 through 6
-        # Soft 14 doubles against dealer 5 through 6
-        # Soft 13 doubles against dealer 5 through 6
-        ace.points = points
-        self.should_double_down(ace, mask)
-
-    should_not_double_down_hard_param = [
-        pytest.param(i, [])
-        for i in range(9, 22)
-        if i not in range(2, 12)
-    ]
-
-    @pytest.mark.parametrize(
-        'points, mask',
-        [
-            (10, range(2, 10)),
-            (11, range(2, 12)),
-            (9, range(3, 7)),
-        ] + should_not_double_down_hard_param
-    )
-    def test_should_double_down_hard(
-        self,
-        mask: list[int],
-        non_ace: Card,
-        points: int,
-    ):
-        """player should double down with hard hand"""
-        # Hard 11 always doubles.
-        # Hard 10 doubles against dealer 2 through 9
-        # Hard 9 doubles against dealer 3 through 6
-        non_ace.points = points
-        self.should_double_down(non_ace, mask)
-
     def test_can_double_down_true(
         self,
         ace: Card,
@@ -150,48 +77,6 @@ class TestDoubleDown():
 
 class TestSplit():
     """test split related functions"""
-    should_not_split_param = [
-        pytest.param(face, [])
-        for face in [Face.TEN, Face.FIVE, Face.JACK, Face.KING, Face.QUEEN]
-    ]
-
-    @pytest.mark.parametrize(
-        'face, mask',
-        [
-            (Face.ACE, range(2, 12)),
-            (Face.EIGHT, range(2, 12)),
-            (Face.FOUR, [5, 6]),
-            (Face.NINE, list(range(2, 7)) + [8, 9]),
-            (Face.SEVEN, range(2, 8)),
-            (Face.SIX, range(2, 7)),
-            (Face.THREE, range(2, 8)),
-            (Face.TWO, range(2, 8)),
-        ] + should_not_split_param
-    )
-    def test_should_split(
-        self,
-        face: Face,
-        mask: list[int],
-        player: BlackjackPlayer,
-    ):
-        """player should split"""
-        # Always split As
-        # 9s splits against dealer 2 through 9, not 7
-        # Always split 8s
-        # 7s splits against dealer 2 through 7
-        # 6s splits against dealer 2 through 6
-        # 4s splits against dealer 5 and 6
-        # 3s splits against dealer 2 through 7
-        # 2s splits against dealer 2 through 7
-        card = Card(face, None)
-        player.add_card(card)
-        player.add_card(card)
-        for i in range(2, 12):
-            if i in mask:
-                assert player.should_split(i)
-            else:
-                assert not player.should_split(i)
-
     def test_can_split_true(
         self,
         ace: Card,
@@ -281,43 +166,6 @@ class TestStand():
         player.add_card(ace)
         assert not player.can_stand()
 
-    def test_should_stand18(
-        self,
-        ace: Card,
-        player: BlackjackPlayer
-    ):
-        """player should not stand with soft 18 against dealer 9 through Ace"""
-        player.add_card(ace)
-        ace.points = 18
-        for i in range(2, 9):
-            assert player.should_stand(i)
-        for i in range(9, 12):
-            assert not player.should_stand(i)
-
-    def test_should_stand(
-        self,
-        non_ace: Card,
-        player: BlackjackPlayer
-    ):
-        """player should stand with 17+"""
-        player.add_card(non_ace)
-        for i in range(17, 22):
-            non_ace.points = i
-            for j in range(9, 12):
-                assert player.should_stand(j)
-
-    def test_should_not_stand(
-        self,
-        non_ace: Card,
-        player: BlackjackPlayer
-    ):
-        """player should not stand with 16-"""
-        player.add_card(non_ace)
-        for i in range(4, 17):
-            non_ace.points = i
-            for j in range(9, 12):
-                assert not player.should_stand(j)
-
     def test_stand(
         self,
         non_ace: Card,
@@ -367,45 +215,6 @@ class TestSurrender():
         player.add_card(ace)
         player.hands.append('some hand')
         assert not player.can_surrender()
-
-    def test_should_surrender15(
-        self,
-        non_ace: Card,
-        player: BlackjackPlayer
-    ):
-        """player should surrender with 15 against dealer 10"""
-        player.add_card(non_ace)
-        non_ace.points = 15
-        assert player.should_surrender(10)
-        for i in range(2, 12):
-            if i == 10:
-                continue
-            assert not player.should_surrender(i)
-
-    def test_should_surrender16(
-        self,
-        non_ace: Card,
-        player: BlackjackPlayer
-    ):
-        """player should surrender with 15 against dealer 9 through Ace"""
-        player.add_card(non_ace)
-        non_ace.points = 16
-        for i in range(9, 12):
-            assert player.should_surrender(10)
-        for i in range(2, 9):
-            assert not player.should_surrender(i)
-
-    def test_should_not_surrender(
-        self,
-        non_ace: Card,
-        player: BlackjackPlayer
-    ):
-        """player should not surrender otherwise"""
-        player.add_card(non_ace)
-        for i in range(4, 22):
-            if i in [15, 16]:
-                continue
-            assert not player.should_surrender(i)
 
     def test_surrender(
         self,
